feat(validation): refactored validation
