+++ backend/domain/use_cases/admin_usecases.py
@@ -1,6 +1,7 @@
-from domain.entities.email import Email
+from tools.validation import email_validation, password_validation
+from tools.func_for_psw import password_hashing
 from domain.entities.admin import Admin
-from domain.value_objects.password import Password
+
 from tools.for_db.work_with_admin_info import add_admin
 
 
@@ -11,8 +12,8 @@ class AdminRegister:
         self.password = password
 
     def admin_register(self):
-        checked_email = Email(self.email)
-        checked_password = Password(self.password)
-        hashed_password = checked_password.get_hashed_password()
-        admin = Admin(checked_email.get_value(), hashed_password)
+        checked_email = email_validation(self.email)
+        checked_password = password_validation(self.password)
+        hashed_password = password_hashing(checked_password)
+        admin = Admin(checked_email, hashed_password)
         add_admin(admin)
