+++ backend/tests/test_func_for_psw.py
@@ -1,4 +1,4 @@
-from domain.value_objects.password import Password
+from tools.func_for_psw import password_hashing
 from tools.func_for_psw import check_psw, get_log_psw_from_header
 
 
@@ -13,8 +13,8 @@ def test_get_psw_from_header():
 def test_prepare_and_check_psw():
     psw = 'Pyth0n'
     psw_2 = 'psw_2'
-    hashed_psw = Password(psw).get_hashed_password()
-    hashed_psw2 = Password(psw_2).get_hashed_password()
+    hashed_psw = password_hashing(psw)
+    hashed_psw2 = password_hashing(psw_2)
     assert check_psw(psw, hashed_psw) is True
     assert check_psw(psw_2, hashed_psw2) is True
     assert check_psw(psw_2, hashed_psw) is False
